+++ pages/events/index.vue
@@ -1,0 +1,14 @@
+<template>
+  <UContainer class="py-12">
+    <h1 class="text-3xl text-center font-bold mb-8 text-primary">Événements</h1>
+    <EventCalendar />
+  </UContainer>
+</template>
+
+<script setup lang="ts">
+import EventCalendar from '@/components/content/EventCalendar.vue'
+</script>
+
+<style>
+
+</style>